#AI commit# CC: 优化向量解析精度处理和类型一致性 / Optimize vector parsing precision and type consistency
## 优化内容 / Optimizations

### 1. 统一浮点数类型为Double
- 将parseSparseVector返回类型从Map<String, Float>改为Map<String, Double>
- 原因：ES对浮点数统一按double处理，保持类型一致性
- 更新相关注释和测试用例

### 2. 优化BigDecimal获取方式
使用JsonNode的decimalValue()方法直接获取BigDecimal：
```java
// 优化前 / Before
new BigDecimal(jsonNode.get(i).asText()).doubleValue()
new BigDecimal(entry.getValue().asText()).floatValue()

// 优化后 / After
jsonNode.get(i).decimalValue().doubleValue()
entry.getValue().decimalValue().doubleValue()
```

优势：
- 代码更简洁高效
- 避免不必要的字符串转换
- 直接利用JsonNode内部的BigDecimal

## 修改文件 / Modified Files
- Elastic8xColumn.java: 优化parseVector和parseSparseVector方法
- Elastic8xFieldDataType.java: 更新SPARSE_VECTOR枚举注释
- Elastic8xColumnTest.java: 更新测试用例类型断言

## 测试结果 / Test Results
Tests run: 23, Failures: 0, Errors: 0, Skipped: 0

Unify floating-point types to Double for consistency with ES's handling of floating-point numbers.
Optimize BigDecimal retrieval using JsonNode's decimalValue() method instead of string conversion.

+++ exchangis-engines/engines/datax/datax-elasticsearch8xwriter/src/test/java/com/webank/wedatasphere/exchangis/datax/plugin/writer/elasticsearchwriter/v8x/column/Elastic8xColumnTest.java
@@ -605,13 +605,13 @@ public class Elastic8xColumnTest {
         assertThat(sparseEmbedding).isInstanceOf(Map.class);
 
         @SuppressWarnings("unchecked")
-        Map<String, Float> sparseVector = (Map<String, Float>) sparseEmbedding;
+        Map<String, Double> sparseVector = (Map<String, Double>) sparseEmbedding;
 
-        // 验证精度：使用BigDecimal作为中间层应该保证精度
-        // Verify precision: using BigDecimal as intermediate should preserve precision
-        assertThat(sparseVector.get("value1")).isEqualTo(0.1f);
-        assertThat(sparseVector.get("value2")).isEqualTo(0.2f);
-        assertThat(sparseVector.get("value3")).isEqualTo(0.3f);
+        // 验证精度：使用decimalValue()直接获取BigDecimal应该保证精度
+        // Verify precision: using decimalValue() to directly get BigDecimal should preserve precision
+        assertThat(sparseVector.get("value1")).isEqualTo(0.1);
+        assertThat(sparseVector.get("value2")).isEqualTo(0.2);
+        assertThat(sparseVector.get("value3")).isEqualTo(0.3);
     }
 
     @Test
@@ -668,11 +668,11 @@ public class Elastic8xColumnTest {
         assertThat(textEmbedding).isInstanceOf(Map.class);
 
         @SuppressWarnings("unchecked")
-        Map<String, Float> sparseVector = (Map<String, Float>) textEmbedding;
+        Map<String, Double> sparseVector = (Map<String, Double>) textEmbedding;
         assertThat(sparseVector)
                 .hasSize(3)
-                .containsEntry("I", 0.55f)
-                .containsEntry("had", 0.4f)
-                .containsEntry("the", 0.3f);
+                .containsEntry("I", 0.55)
+                .containsEntry("had", 0.4)
+                .containsEntry("the", 0.3);
     }
 }
